requirements.txt
kafka-python
boto3
snowflake-connector-python>=3.10.0
dbt-core
dbt-snowflake
requests